+++ src/main/java/com/publicuhc/pluginframework/routing/DefaultCommandRoute.java
@@ -1,6 +1,7 @@
 package com.publicuhc.pluginframework.routing;
 
 import com.publicuhc.pluginframework.routing.exception.CommandInvocationException;
+import com.publicuhc.pluginframework.routing.proxy.MethodProxy;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.craftbukkit.libs.joptsimple.OptionException;
